backend/requirements.txt
django==5.2.9
djangorestframework==3.16.1
django-cors-headers==4.9.0
python-dotenv==1.2.1
psycopg2-binary==2.9.11
dj-database-url==3.1.0
djangorestframework_simplejwt==5.5.1
groq
